Add files via upload
This repository is part of a lesson using the Stanford open policing data set. Since my students are taking a first course in python or have experience in python, I did this as a Jupyter notebook. I'd recommend that you take your own city's data. And if you can't get that data, take it to the streets.

-Joel